compile
git-svn-id: https://nlg0.isi.edu/svn/sbmt/trunk/graehl@3541 7f356e55-ce0a-0410-be72-bc6dfe96aa1c

+++ shared/command_line.hpp
@@ -43,7 +43,7 @@ struct argc_argv : private std::stringbuf
     std::vector<char const*> argvptrs;
     int argc() const
     {
-        return argvptrs.size();
+      return (int)argvptrs.size();
     }
     argv_t argv() const
     {
